+++ nextgisweb/feature_layer/migration/438055b2-add-text_search-column.rw.sql
@@ -1,0 +1,3 @@
+/*** { "revision": "438055b2" } ***/
+
+ALTER TABLE layer_field DROP COLUMN text_search;