+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changes
 - Configurable permissions for groups, users, SRS and CORS management.
 - Lots of customization options for embedded webmaps: map tools and panels.
 - Support for custom SVG logos in addition to PNG.
+- Ability to exclude feature layer fields from text search.
 
 
 4.6.0